# docs/swagger.yaml
definitions:
  dto.Customer:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.CustomerInput:
    properties:
      name:
        type: string
    type: object
  dto.Deposit:
    properties:
      amount:
        type: number
      created_at:
        type: string
      customer_id:
        type: string
      fund_id:
        type: string
      id:
        type: string
    type: object
  dto.DepositInput:
    properties:
      amount:
        type: number
      customer_id:
        type: string
      fund_id:
        type: string
    type: object
  dto.Fund:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.FundInput:
    properties:
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /customer:
    post:
      description: Add a customer
      parameters:
      - description: Customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/dto.CustomerInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Customer'
      summary: Add a customer
      tags:
      - Customers
  /customer/{customer_id}/deposits:
    get:
      description: deposits for a specific customer
      parameters:
      - description: Customer ID
        in: path
        name: customer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Deposit'
            type: array
      summary: Get deposits for a specific customer
      tags:
      - Customers
  /customer/list:
    get:
      consumes:
      - application/json
      description: Get the list of customers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Customer'
            type: array
      summary: Get the list of customers
      tags:
      - Customers
  /deposit:
    post:
      description: Add a deposit
      parameters:
      - description: Deposit data
        in: body
        name: deposit
        required: true
        schema:
          $ref: '#/definitions/dto.DepositInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Deposit'
      summary: Add a deposit
      tags:
      - Deposits
  /fund:
    post:
      description: Add a fund
      parameters:
      - description: Fund data
        in: body
        name: fund
        required: true
        schema:
          $ref: '#/definitions/dto.FundInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Fund'
      summary: Add a fund
      tags:
      - Funds
  /fund/list:
    get:
      consumes:
      - application/json
      description: Get the list of funds
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Fund'
            type: array
      summary: Get the list of funds
      tags:
      - Funds
swagger: "2.0"